common: Introduce the common workspace to work with the socket also in other craters
