+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
 members = [
         "rpc",
-        "plugin",
+        "common",
 ]
