+++ Makefile
@@ -1,0 +1,13 @@
+CC=cargo
+FMT=fmt
+
+OPTIONS=
+
+default: fmt
+	$(CC) build
+
+fmt:
+	$(CC) fmt --all
+
+check:
+	$(CC) test --all
