+++ common/src/lib.rs
@@ -1,0 +1,10 @@
+/// Common crate that implements all the primitive
+/// to talk the with the socket
+///
+/// author: https://github.com/vincenzopalazzo
+extern crate serde;
+extern crate serde_json;
+
+pub mod client;
+pub mod errors;
+pub mod types;
