+++ rpc/Cargo.toml
@@ -14,6 +14,7 @@ edition = "2018"
 serde = "1"
 serde_derive = "1"
 serde_json = "1.0"
+common = { path = "../common" }
 
 [dev-dependencies]
 rstest = "0.10.0"
