+++ rpc/src/error.rs
@@ -12,16 +12,6 @@
 // If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 //
 
-//! Error handling
-//!
-//! Some useful methods for creating Error objects
-//!
-
-use std::io;
-use std::{error, fmt};
-
-use serde_json;
-
 /// Known lightningd error codes.
 ///
 /// **Keep this up to date with `lightningd/jsonrpc_errors.h`**
@@ -79,72 +69,4 @@ pub enum RpcErrorCode {
     INVOICE_LABEL_ALREADY_EXISTS = 900,
     /// Invoice pre-image already exists
     INVOICE_PREIMAGE_ALREADY_EXISTS = 901,
-}
-
-/// A library error
-#[derive(Debug)]
-pub enum Error {
-    /// Json error
-    Json(serde_json::Error),
-    /// IO Error
-    Io(io::Error),
-    /// Error response
-    Rpc(RpcError),
-    /// Response has neither error nor result
-    NoErrorOrResult,
-    /// Response to a request did not have the expected nonce
-    NonceMismatch,
-    /// Response to a request had a jsonrpc field other than "2.0"
-    VersionMismatch,
-}
-
-impl From<serde_json::Error> for Error {
-    fn from(e: serde_json::Error) -> Error {
-        Error::Json(e)
-    }
-}
-
-impl From<io::Error> for Error {
-    fn from(e: io::Error) -> Error {
-        Error::Io(e)
-    }
-}
-
-impl From<RpcError> for Error {
-    fn from(e: RpcError) -> Error {
-        Error::Rpc(e)
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            Error::Json(ref e) => write!(f, "JSON decode error: {}", e),
-            Error::Io(ref e) => write!(f, "IO error response: {}", e),
-            Error::Rpc(ref r) => write!(f, "RPC error response: {:?}", r),
-            Error::NoErrorOrResult => write!(f, "Malformed RPC response"),
-            Error::NonceMismatch => write!(f, "Nonce of response did not match nonce of request"),
-            Error::VersionMismatch => write!(f, "`jsonrpc` field set to non-\"2.0\""),
-        }
-    }
-}
-
-impl error::Error for Error {
-    fn cause(&self) -> Option<&dyn error::Error> {
-        match *self {
-            Error::Json(ref e) => Some(e),
-            _ => None,
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
-/// A JSONRPC error object
-pub struct RpcError {
-    /// The integer identifier of the error
-    pub code: i32,
-    /// A string describing the error
-    pub message: String,
-    /// Additional data specific to the error
-    pub data: Option<serde_json::Value>,
 }
